Add remote-only upload

+++ daman/services/aws.py
@@ -1,10 +1,12 @@
 import boto3
 from pathlib import Path
-from warnings import warn
-from os.path import getsize
+from logging import getLogger
 from typing import Union, IO, AnyStr
 
 from daman.services import Provider, Progress
+
+
+logger = getLogger(__name__)
 
 
 class AWSProvider(Provider):
@@ -29,7 +31,7 @@ class AWSProvider(Provider):
             )
         elif buffer is not None:
             self.s3.meta.client.download_fileobj(
-                Bucket=self.bucket, Key=key, Fileobj=buffer, Callback=pbar,
+                Fileobj=buffer, Key=key, Bucket=self.bucket, Callback=pbar,
             )
         else:
             raise ValueError("`buffer` or `file_path` cannot be `None` simultaneously.")
@@ -37,26 +39,49 @@ class AWSProvider(Provider):
         # close progress bar
         pbar.close()
 
-    def upload(self, key: str, file_path: Union[str, Path]):
-        msg = f"{file_path} does not exist."
-        assert file_path.exists(), msg
-        # Initialising progress bar
-        pbar = Progress(size=getsize(file_path), key=key)
+    def upload(
+        self, key: str, file_path: Union[str, Path] = None, buffer: IO[AnyStr] = None
+    ):
+        if file_path is not None:
+            msg = f"{file_path} does not exist."
+            assert file_path.exists(), msg
+            # Initialising progress bar
+            pbar = Progress(size=20, key=key)
 
-        # Upload to bucket
-        self.s3.meta.client.upload_file(
-            Bucket=self.bucket,
-            Key=key,
-            Filename=str(file_path),
-            Callback=pbar,
-            ExtraArgs={"Metadata": {"md5": self.md5(file_path=file_path)}},
-        )
+            # Upload to bucket
+            self.s3.meta.client.upload_file(
+                Bucket=self.bucket,
+                Key=key,
+                Filename=str(file_path),
+                Callback=pbar,
+                ExtraArgs={"Metadata": {"md5": self.md5(file_path=file_path)}},
+            )
 
-        # close progres bar
-        pbar.close()
+            # close progres bar
+            pbar.close()
+        elif buffer is not None:
+            # Initialising progress bar
+            pbar = Progress(size=buffer.getbuffer().nbytes, key=key)
+            # Upload to bucket
+            self.s3.meta.client.upload_fileobj(
+                Fileobj=buffer,
+                Key=key,
+                Bucket=self.bucket,
+                Callback=pbar,
+                ExtraArgs={"Metadata": {"md5": self.md5(buffer=buffer)}},
+            )
+
+            # close progres bar
+            pbar.close()
+        else:
+            msg = "either `file_path` or `buffer` must be provided."
+            raise ValueError(msg)
 
     def delete(self, key: str):
-        self.s3.meta.client.delete_object(Bucket=self.bucket, Key=key)
+        if key in self.keys:
+            self.s3.meta.client.delete_object(Bucket=self.bucket, Key=key)
+        else:
+            logger.warning(f"`{key}` already deleted from cloud service.")
 
     @property
     def keys(self):
@@ -76,5 +101,5 @@ class AWSProvider(Provider):
             remote_md5 = self.s3.Bucket(self.bucket).Object(key).metadata["md5"]
             return local_md5 == remote_md5
         else:
-            warn(msg)
+            logger.warning(msg)
             return True
